Move JoinRoomForm to its own uncontrolled component
